Refactor backup system by consolidating code editor functionality
- Removed individual modules for Antigravity and VS Code backup implementations.
- Introduced a unified `code_editors.rs` module to handle backup for multiple code editors.
- Implemented a new `CodeEditorCli` struct to manage extensions and settings for both VS Code and Antigravity.
- Updated the `system.rs` module to integrate the new backup structure.
- Removed obsolete files and traits related to the previous backup implementations.
- Ensured compatibility with existing functionality while streamlining the codebase.
