+++ src/app/backup/run.rs
@@ -38,8 +38,9 @@ pub fn execute(ctx: &AppContext, component_input: &str) -> Result<(), AppError> 
             let output_file = local_config_dir.join("system.yml");
             backup::system::execute(ctx, &definitions_dir, &output_file)
         }
-        BackupComponent::Vscode => backup::vscode::execute(ctx, &local_config_dir),
-        BackupComponent::Antigravity => backup::antigravity::execute(ctx, &local_config_dir),
+        BackupComponent::Vscode | BackupComponent::Antigravity => {
+            backup::code_editors::execute(ctx, component, &local_config_dir)
+        }
     }?;
 
     println!();
